refactor: mv kubeconn -> kubeutil pkg
